changed floats to doubles
Former-commit-id: 7b390a9e06fd7057679ad224be9fa139e6ef5adb

+++ drone_gnc/src/models/drone_model.hpp
@@ -8,11 +8,11 @@
 class Drone : public Rocket {
 public:
 
-    float minPropellerSpeed;
-    float maxPropellerSpeed;
+    double minPropellerSpeed;
+    double maxPropellerSpeed;
 
-    float maxServo1Angle;
-    float maxServo2Angle;
+    double maxServo1Angle;
+    double maxServo2Angle;
 
     double thrust_scaling;
     Eigen::Vector3d disturbance_torque;
